refactor: split `KeyboardMovementObserver` implementation
